MB-36923: Allow "default" and "disk_io_optimized" for ThreadCount
Allow the values "default" and "disk_io_optimized" to be used for the
num_reader_threads & num_writer_threads memcached.json settings:

* "default": Set the number of reader/writer threads to be a balanced
  value which is reasonable for most workloads:

  - num_writers: 4
  - num_readers: Number of available CPU cores; minimum 4, maximum 16.

  This is broadly similar to the 6.0 default calculated values.

* "disk_io_optimized": Set the number of reader/writer threads to be
  optimised for Disk IO latency:

  - num_writers: Number of available CPU cores; minimum 4, maximum 64.
  - num_readers: Number of available CPU cores; minimum 4, maximum 64.

Reviewed-on: http://review.couchbase.org/118511
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>

+++ daemon/settings.h
@@ -823,11 +823,11 @@ public:
         notify_changed("parent_identifier");
     }
 
-    size_t getNumReaderThreads() const {
+    int getNumReaderThreads() const {
         return num_reader_threads.load(std::memory_order_acquire);
     }
 
-    void setNumReaderThreads(size_t val) {
+    void setNumReaderThreads(int val) {
         num_reader_threads.store(val, std::memory_order_release);
         has.num_reader_threads = true;
         notify_changed("num_reader_threads");
@@ -1024,8 +1024,8 @@ protected:
     /// The identifier to use for the parent monitor
     int parent_identifier = -1;
 
-    std::atomic<size_t> num_reader_threads{0};
-    std::atomic<size_t> num_writer_threads{0};
+    std::atomic<int> num_reader_threads{0};
+    std::atomic<int> num_writer_threads{0};
 
 public:
     /**
